+++ wcmatch/pathlib.py
@@ -11,7 +11,8 @@ __all__ = (
     "REALPATH", "FOLLOW", "MATCHBASE", "NEGATEALL", "NODIR",
     "C", "F", "I", "R", "D", "E", "N", "B", "M", "P", "L", "S", "X", "O", "A",
     "escape", "raw_escape",
-    "Path", "PurePath", "WindowsPath", "PosixPath", "PurePosixPath", "PureWindowsPath"
+    "Path", "PurePath", "WindowsPath", "PosixPath", "PurePosixPath", "PureWindowsPath",
+    "AUTO", "WINDOWS", "UNIX"
 )
 
 C = CASE = glob.CASE
@@ -29,6 +30,10 @@ S = SPLIT = glob.SPLIT
 X = MATCHBASE = glob.MATCHBASE
 O = NODIR = glob.NODIR
 A = NEGATEALL = glob.NEGATEALL
+
+AUTO = glob.AUTO
+WINDOWS = glob.WINDOWS
+UNIX = glob.UNIX
 
 FLAG_MASK = (
     CASE |
@@ -50,9 +55,6 @@ FLAG_MASK = (
     _wcparse.GLOBSTAR |
     _wcparse.PATHNAME
 )
-
-raw_escape = glob.raw_escape
-escape = glob.escape
 
 
 class Path(pathlib.Path):
@@ -189,3 +191,15 @@ class WindowsPath(Path, PureWindowsPath):
     """Windows path."""
 
     __slots__ = ()
+
+
+def raw_escape(pattern, *, platform=AUTO):
+    """Apply raw character transform before applying escape."""
+
+    return glob.raw_escape(pattern, platform=platform)
+
+
+def escape(pattern, *, platform=AUTO):
+    """Escape."""
+
+    return glob.escape(pattern, platform=platform)